Refactor cooking and code sections
